[REBORN] Implemented Per-Ability and Item quickcast/normal cast toggles
Using Alt+Space, and an ability / item key, that ability / item is
toggled from quick-cast to normal-cast and back individually. "F7" will
force all abilities and items to either quick-cast or normal-cast
depending on how many times it is pressed.

NOTE: When Alt+Space is pressed, both keys have to be released before
either Alt or Space can be used as modifiers again.